Create Searching a number
Search in a Rotated Array
Given a sorted and rotated array A of N distinct elements which is rotated at some point, and given an element key. The task is to find the index of the given element key in the array A.
Expected Time Complexity: O(log N).
Expected Auxiliary Space: O(1).